Refactor: change structure
