+++ .goreleaser.yaml
@@ -9,7 +9,7 @@ before:
 builds:
   - env:
       - CGO_ENABLED=0
-    main: ./cmd/
+    main: ./cmd/hln/
     binary: hln
     ldflags:
       - -s -w
